requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.6.0
click==8.1.3
colorama==0.4.5
Django==4.0.6
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.2
